+++ exadata-database/outputs.tf
@@ -13,19 +13,19 @@ output "cloud_vm_clusters" {
 
 output "database_homes" {
   description = "The deployed Databases Homes in the OCI Database Service."
-  value       = var.enable_output ? oci_database_db_home.these : null
+  value       = var.enable_output ? module.common_database.database_homes : null
   sensitive   = true
 }
 
 output "databases" {
   description = "The deployed Databases in the OCI Database Service."
-  value       = var.enable_output ? oci_database_database.these : null
+  value       = var.enable_output ? module.common_database.databases : null
   sensitive   = true
 }
 
 output "pluggable_databases" {
   description = "The deployed Pluggable Databases in the OCI Database Service."
-  value       = var.enable_output ? oci_database_pluggable_database.these : null
+  value       = var.enable_output ? module.common_database.pluggable_databases : null
   sensitive   = true
 }
 
@@ -33,9 +33,9 @@ locals {
   cloud_exadata_database_resources = {
     cloud_exadata_infrastructures = { for k, v in oci_database_cloud_exadata_infrastructure.these : k => { "id" : v.id, "compartment_id" : v.compartment_id } }
     cloud_vm_clusters             = { for k, v in oci_database_cloud_vm_cluster.these : k => { "id" : v.id, "compartment_id" : v.compartment_id } }
-    database_homes                = { for k, v in oci_database_db_home.these : k => { "id" : v.id, "compartment_id" : try(v.compartment_id, null) } }
-    databases                     = { for k, v in oci_database_database.these : k => { "id" : v.id } }
-    pluggable_databases           = { for k, v in oci_database_pluggable_database.these : k => { "id" : v.id } }
+    database_homes                = try(module.common_database.database_resources.database_homes, {})
+    databases                     = try(module.common_database.database_resources.databases, {})
+    pluggable_databases           = try(module.common_database.database_resources.pluggable_databases, {})
   }
 }